📚minimal works cli program
