+++ packages/curvenote/src/utils/index.ts
@@ -1,1 +1,3 @@
 export * from './utils.js';
+export * from './web.js';
+export * from './types.js';
