+++ packages/curvenote/src/web/deploy.ts
@@ -1,244 +1,17 @@
-import { createHash } from 'node:crypto';
 import chalk from 'chalk';
-import cliProgress from 'cli-progress';
-import fs from 'node:fs';
-import mime from 'mime-types';
 import { selectors, buildSite, clean } from 'myst-cli';
 import { tic } from 'myst-cli-utils';
-import type { Logger } from 'myst-cli-utils';
-import fetch from 'node-fetch';
-import path from 'node:path';
-import pLimit from 'p-limit';
-import type {
-  DnsRouter,
-  SiteDeployRequest,
-  SiteUploadRequest,
-  SiteUploadResponse,
-} from '@curvenote/blocks';
+import type { DnsRouter } from '@curvenote/blocks';
 import { MyUser } from '../models.js';
 import type { ISession } from '../session/types.js';
-import { addOxaTransformersToOpts, confirmOrExit } from '../utils/index.js';
+import {
+  addOxaTransformersToOpts,
+  confirmOrExit,
+  uploadContentAndDeployToPrivateCdn,
+  uploadContentAndDeployToPublicCdn,
+} from '../utils/index.js';
 import type { SiteConfig } from 'myst-config';
-import { getHeaders } from '../session/tokens.js';
-
-type FromTo = {
-  from: string;
-  to: string;
-};
-
-function listFolderContents(session: ISession, from: string, to = ''): FromTo[] {
-  const directory = fs.readdirSync(from);
-  const files: string[] = [];
-  const folders: string[] = [];
-  directory.forEach((f) => {
-    if (fs.statSync(path.join(from, f)).isDirectory()) {
-      folders.push(f);
-    } else {
-      files.push(f);
-    }
-  });
-  const outputFiles: FromTo[] = files.map((f) => ({
-    from: path.join(from, f),
-    to: to ? `${to}/${f}` : f,
-  }));
-  const outputFolders: FromTo[] = folders
-    .map((f) => listFolderContents(session, path.join(from, f), to ? `${to}/${f}` : f))
-    .flat();
-  return [...outputFiles, ...outputFolders];
-}
-
-async function prepareFileForUpload(from: string, to: string): Promise<FileInfo> {
-  const content = fs.readFileSync(from).toString();
-  const stats = fs.statSync(from);
-  const md5 = createHash('md5').update(content).digest('hex');
-  const contentType = mime.lookup(path.extname(from));
-  return { from, to, md5, size: stats.size, contentType: contentType || '' };
-}
-
-type FileInfo = {
-  from: string;
-  to: string;
-  md5: string;
-  size: number;
-  contentType: string;
-};
-
-type FileUpload = FileInfo & {
-  bucket: string;
-  signedUrl: string;
-};
-
-async function uploadFile(log: Logger, upload: FileUpload) {
-  const toc = tic();
-  log.debug(`Starting upload of ${upload.from}`);
-  const resumableSession = await fetch(upload.signedUrl, {
-    method: 'POST',
-    headers: {
-      'x-goog-resumable': 'start',
-      'content-type': upload.contentType,
-    },
-  });
-  // Endpoint to which we should upload the file
-  const location = resumableSession.headers.get('location') as string;
-
-  // we are not resuming! if we want resumable uploads we need to implement
-  // or use something other than fetch here that supports resuming
-  const readStream = fs.createReadStream(upload.from);
-  const uploadResponse = await fetch(location, {
-    method: 'PUT',
-    headers: {
-      'Content-length': `${upload.size}`,
-    },
-    body: readStream,
-  });
-
-  if (!uploadResponse.ok) {
-    log.error(`Upload failed for ${upload.from}`);
-  }
-
-  log.debug(toc(`Finished upload of ${upload.from} in %s.`));
-}
-
-export async function prepareUploadRequest(session: ISession) {
-  const filesToUpload = listFolderContents(session, session.sitePath());
-  session.log.info(`🔬 Preparing to upload ${filesToUpload.length} files`);
-
-  const files = await Promise.all(
-    filesToUpload.map(({ from, to }) => prepareFileForUpload(from, to)),
-  );
-
-  const uploadRequest: SiteUploadRequest = {
-    files: files.map(({ md5, size, contentType, to }) => ({
-      path: to,
-      content_type: contentType,
-      md5,
-      size,
-    })),
-  };
-
-  return { files, uploadRequest };
-}
-
-export async function processUpload(
-  session: ISession,
-  files: FileInfo[],
-  uploadTargets: SiteUploadResponse,
-  opts?: { ci?: boolean },
-) {
-  // Only upload N files at a time
-  const limit = pLimit(10);
-  const bar1 = opts?.ci
-    ? undefined
-    : new cliProgress.SingleBar({}, cliProgress.Presets.shades_classic);
-  session.log.info(`☁️  Uploading ${files.length} files`);
-  bar1?.start(files.length, 0);
-  const toc = tic();
-  let current = 0;
-  await Promise.all(
-    files.map((file) =>
-      limit(async () => {
-        const upload = uploadTargets.files[file.to];
-        await uploadFile(session.log, {
-          bucket: uploadTargets.bucket,
-          from: file.from,
-          to: upload.path,
-          md5: file.md5,
-          size: file.size,
-          contentType: file.contentType,
-          signedUrl: upload.signed_url,
-        });
-        current += 1;
-        bar1?.update(current);
-        if (opts?.ci && current % 5 == 0) {
-          session.log.info(`☁️  Uploaded ${current} / ${files.length} files`);
-        }
-      }),
-    ),
-  );
-  bar1?.stop();
-  session.log.info(toc(`☁️  Uploaded ${files.length} files in %s.`));
-
-  const cdnKey = uploadTargets.id;
-  return { cdnKey };
-}
-
-/**
- * Upload content to the (private) staging bucket
- *
- * @param session
- * @param opts
- * @returns cdnkey and filepaths for deployment
- */
-async function uploadContent(session: ISession, opts?: { ci?: boolean }) {
-  const { files, uploadRequest } = await prepareUploadRequest(session);
-  const { json: uploadTargets } = await session.post<SiteUploadResponse>('/sites/upload', {
-    ...uploadRequest,
-  });
-  const { cdnKey } = await processUpload(session, files, uploadTargets, opts);
-  return { cdnKey, filepaths: files.map(({ to }) => ({ path: to })) };
-}
-
-/**
- * Deploy content to the public/private CDN based on the usePublicCdn argument
- *
- * @param session
- * @param usePublicCdn whether to deploy to the public CDN or default private CDN
- * @param cdnKey
- * @param filepaths
- * @returns cdnkey
- */
-async function deployContent(
-  session: ISession,
-  privacy: { public: boolean },
-  cdnKey: string,
-  filepaths: { path: string }[],
-) {
-  const toc = tic();
-  const deployRequest: SiteDeployRequest = {
-    public: privacy.public,
-    id: cdnKey,
-    files: filepaths,
-  };
-  const deployResp = await session.post('/sites/deploy', deployRequest);
-
-  if (deployResp.ok) {
-    session.log.info(toc(`🚀 Deployed ${filepaths.length} files in %s.`));
-    session.log.debug(`CDN key: ${cdnKey}`);
-  } else {
-    throw new Error('Deployment failed: Please contact support@curvenote.com!');
-  }
-  return cdnKey;
-}
-
-/**
- * Perform PRIVATE CDN upload and deployment
- *
- * @param session
- * @param opts
- * @returns cdnkey
- */
-export async function uploadContentAndDeployToPrivateCdn(
-  session: ISession,
-  opts?: { ci?: boolean },
-) {
-  const { cdnKey, filepaths } = await uploadContent(session, opts);
-  return await deployContent(session, { public: false }, cdnKey, filepaths);
-}
-
-/**
- * Perform PUBLIC CDN upload and deployment
- *
- * @param session
- * @param opts
- * @returns cdnkey
- */
-export async function uploadContentAndDeployToPublicCdn(
-  session: ISession,
-  opts?: { ci?: boolean },
-) {
-  const { cdnKey, filepaths } = await uploadContent(session, opts);
-  return await deployContent(session, { public: true }, cdnKey, filepaths);
-}
+import { postNewWork, submitToVenue } from '../works/utils.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export async function preflightPromotePublicContent(session: ISession, domains?: string[]) {
@@ -254,90 +27,6 @@ export async function preflightPromoteToVenue(
   domains?: string[],
 ) {
   // TODO throw on no permission to submit to venue
-}
-
-interface WorkBody {
-  id: string;
-  cdn: string;
-}
-
-interface SubmissionBody {
-  work_version_id: string;
-  kind: string;
-}
-
-async function postToJournals(
-  session: ISession,
-  pathname: string,
-  body: WorkBody | SubmissionBody,
-) {
-  const apiUrl = `https://journals.curvenote.dev/v1/`;
-  // const apiUrl = `http://localhost:3031/v1/`;
-
-  const url = `${apiUrl}${pathname}`;
-  console.debug('Posting to', url);
-
-  const method = 'POST';
-  const headers = await getHeaders(session.log, (session as any).$tokens);
-  const response = await fetch(url, {
-    method,
-    body: JSON.stringify(body),
-    headers: {
-      'Content-Type': 'application/json',
-      ...headers,
-    },
-  });
-  const json = (await response.json()) as any;
-  if (!response.ok) {
-    const dataString = JSON.stringify(json, null, 2);
-    session.log.debug(`${method.toUpperCase()} FAILED ${url}: ${response.status}\n\n${dataString}`);
-  }
-  return {
-    ok: response.ok,
-    status: response.status,
-    json,
-  };
-}
-
-export async function postNewWork(session: ISession, cdnKey: string, cdn: string) {
-  const toc = tic();
-
-  const resp = await postToJournals(session, 'works', { id: cdnKey, cdn });
-
-  if (resp.ok) {
-    session.log.info(toc(`🚀 Submitted a new work in %s.`));
-    session.log.debug(`CDN key: ${cdnKey}`);
-    session.log.debug(`Work Id: ${resp.json.id}`);
-    session.log.debug(`Work Version Id: ${resp.json.version_id}`);
-    return {
-      cdnKey,
-      workId: resp.json.id,
-      workVersionId: resp.json.version_id,
-    };
-  } else {
-    throw new Error('Posting new work failed: Please contact support@curvenote.com');
-  }
-}
-
-export async function submitToVenue(
-  session: ISession,
-  venue: string,
-  work_version_id: string,
-  kind: string,
-) {
-  const toc = tic();
-  const submissionRequest: SubmissionBody = { work_version_id, kind };
-  const resp = await postToJournals(session, `sites/${venue}/submissions`, submissionRequest);
-  if (resp.ok) {
-    session.log.info(toc(`🚀 Submitted to venue "${venue}" in %s.`));
-    session.log.debug(`Submission id: ${resp.json.id}`);
-    session.log.debug(`Submitted by: ${resp.json.submitted_by.name ?? resp.json.submitted_by.id}`);
-    return {
-      submissionId: resp.json.id,
-    };
-  } else {
-    throw new Error('Submission failed: Please contact support@curvenote.com');
-  }
 }
 
 export async function promotePublicContent(session: ISession, cdnKey: string, domains?: string[]) {
