+++ packages/curvenote/src/works/index.ts
@@ -1,0 +1,3 @@
+export { create } from './create.js';
+export { list } from './list.js';
+export { submit } from './submit.js';
