+++ packages/curvenote/src/works/list.ts
@@ -1,0 +1,1 @@
+export async function list() {}
